Allow system, global and local config, similar to git configs
Viable config file paths are:
* System
  * `/etc/default/adr-config`
  * `/etc/adr-config`
* Global (user)
  * `~/.adr-config`
* Local (repo)
  * `$(_adr_dir)/.adr-config`

This change relies on `_adr_dir` not trying to load config, so
it also incorporates this removal, even though this may have been
removed in another branch/PR.

Tests are included. If it's not possible to run sudo based tests in
the CI/CD system, then these should be removed.